# specs/004-dashboard-task-crud/contracts/tasks.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dashboard & Task CRUD API",
    "description": "API contract for secure task creation, filtering, retrieval, general updates, and status mutations.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/tasks": {
      "post": {
        "summary": "Create a new task",
        "description": "Creates a task linked with the authenticated user.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      },
      "get": {
        "summary": "Retrieve user tasks",
        "description": "Fetches all tasks belonging strictly to the authenticated user. Supports filtering by priority query.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["epic", "high", "medium", "low"]
            },
            "description": "Filter tasks by priority"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
    "/tasks/{id}": {
      "patch": {
        "summary": "Update task details",
        "description": "Updates general fields (description, priority). The title field is immutable and ignored.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/status": {
      "patch": {
        "summary": "Update task status",
        "description": "Dedicated endpoint updating strictly the task's execution state.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "in-progress", "completed"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateTaskRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Implement unit tests"
          },
          "description": {
            "type": "string",
            "example": "Verify dashboard features"
          },
          "priority": {
            "type": "string",
            "enum": ["epic", "high", "medium", "low"],
            "default": "medium",
            "example": "high"
          }
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Updated details"
          },
          "priority": {
            "type": "string",
            "enum": ["epic", "high", "medium", "low"],
            "example": "low"
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65b5974c5d5e5e4078cb8db3"
          },
          "authId": {
            "type": "string",
            "example": "65b5974c5d5e5e4078cb8db1"
          },
          "title": {
            "type": "string",
            "example": "Implement unit tests"
          },
          "description": {
            "type": "string",
            "example": "Verify dashboard features"
          },
          "priority": {
            "type": "string",
            "example": "high"
          },
          "status": {
            "type": "string",
            "example": "pending"
          }
        }
      }
    }
  }
}